project/requirements.txt
Django==5.2.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0

gunicorn
whitenoise
dj-database-url
python-dotenv

numpy
scikit-learn
scipy
torch
sentence-transformers
transformers
pypdf
joblib